Wading into the multi-package/module quagmire.
We're currently using Lerna and some hackery to avoid having to create "index"
modules for our sub-packages. Even though everyone in the JS world loves index
modules, they are problematic for our use case.

We need fine-grained imports directly to modules residing in other libraries so
that we can maintain a fine-grained dependency graph for code reloading. If
everything was imported via a member in a project-spanning uber index module,
then every time anything in the entire library changed everything that depended
on anything in that library would have to be reloaded. Double plus ungood.

So of course, once again, we find ourselves swimming upstream against the river
of shit that is the JavaScript ecosystem. Yay.