Use a workaround for literal type inference.
We now enforce that the reactive primitives are only used on `Data` types,
which is technically necessary because those are the only types for which we
can guarantee sane equality checking and thus correct 'only emit events when
values actually change' semantics.

Furthermore, for types that go over the network, we *really* need them to only
be Data. We can't serialize arbitrary JavaScript objects.

However, I'm sure that special snowflake cases will eventually arise where
someone wants to use local reactive plumbing for local special snowflake types,
so we may eventually want some way for a value to provide its own equality
checking. That will undoubtedly be an ugly hack, but such is life when building
on the shifty sands of JavaScript.