Go back to combining everything into a single library.
Getting Lerna, Jest, and TypeScript all playing nicely together, with the
additional complication that we don't want to use index files, but rather want
to import directly from sub-modules is already turning into a major fiasco.

Rather than take this punch in velocity from the outset, we'll put our faith in
the efficacy of tree shaking (so someone can use tfw-platform just for one or
two sub-libraries and not unnecessarily bloat their game) and that TypeScript
will benefit from our fine-grained import regime and not have to retypecheck
50kLOC every time anything changes.

If necessary in the future, we can consider chunkier breakups (splitting
everything into two or three libraries, but not a dozen), or other mitigations.