Move config module into core.
The "core" of the config system is pretty simple and will be used by nearly
everything if not everything, so there's not much value in separating it from
core.

Further, the UI part of the config library will need to be separate because it
will have all sorts of dependencies that we don't want to foist onto libraries
that just want to make use of config objects. So we'll likely only have a
tfw.configui library which itself depends on ui, scene2 and core.