Rejiggering to enable non-data streams/values.
Ultimately we need streams and values that contain non-value types, so we just
have to bite the bullet and expose some equality testing machinery for those
cases. In general we'll either want data (structural) equality or reference
equality, but I also exposed a mechanism for supplying arbitrary equality test
functions since we need it under the hood anyway.

I kept the "natural" names operating on `Data` values. If you, for example,
unthinkingly call v.map() on a non-Data value, you'll get a compiler error
because the type won't conform to `Data`. Thus we can continue to use
everything "comfortably" for all `T extends Data`, and when you want to create
reactive dataflow for non-Data values, you have to tack on `Ref` here and
there.