Main copy of package.json lives in lib.
src has a symlink to that because we have to build from src into lib so that we
end up with a project that can be grokked by other tools like tsc, webpack,
etc. and also admits fine-grained importing of modules (meaning everything has
to be rooted at the directory that contains package.json, no src or lib dir
allowed).

The crack smoking runs so deep here in NPM land.