Add UniformQuadBatch.
This passes quad data in shader uniforms which allows us to avoid much of the
data duplication that goes with TriangleBatch when drawing textured quads. We
can only render about ~350 quads at a time, but that still turns out to be 25%
faster (6ms avg versus 8ms avg) on our 10 kilobird test.