Have the demo fns start/end the surface/batch.
And also clear if they want to, or not.

I think eventually we either want to move Renderer to core/gl and have it not
do anything scene2-specific, or for the demos, have the scene3 demos use
totally separate GL context and canvas. Probably the latter, since I assume
three.js handles binding the default framebuffer and most everything that
Renderer is doing for scene2.