Use Subject to manage cleanup.
One of Subject's main reasons for existence is to allow the acquisition of
resources upon first subscription and then the relinquishing of those resources
when the last subscriber goes away.

This will need to be better documented because it's not obvious how all the
wiring works. Subject.derive is passed a "setup" function, the setup function
is called the first time a subscriber shows up. It is passed a "dispatch"
function and returns a cleanup function. The body of the function (and any
asynchronous stuff it sets up itself) can call the dispatch function once a
value is known for the subject (and again if it ever changes), and the setup
function returns a cleanup function which will be called when the last
subscriber is removed from the subject. So with a single (nested function) API,
we get setup/change/cleanup.