Host manages collection of Roots.
Bridges the gap (especially with Host2/Host3) more nicely between UI framework
and whatever rendering the game is actually doing. Will also be needed for
free-floating popups &c where a Root will need to be able to add a new root,
cause it to be rendered and then later dispose it.