Factor text measurement & rendering into Span.
This makes it easier to reuse in other UI elements (like the forthcoming
editable Text element), and it will make it easier to add Block later which
will render multi-line and potentially multi-style text.

I'm punting a little on text measurement for now. We can cheat for now with
align-top rendering and the fact that the font size is traditionally the height
of the em block. When we do multi-style text, we'll have to align things to the
proper baseline and come up with more involved workarounds for Firefox's lack
of civilized TextMetrics.