Get TypeDoc working again.
Except not completely because it uses an older version of TypeScript (which we
cannot override because TypeDoc uses sneaky internals and apparently must be
tightly bound to a specific compiler version; and they haven't shipped their
update to TS 3.5.x yet, which we rely on). Sigh.

Here's a tip from the goddamned pros: if you make a fucking computer language,
build the fucking documentation into the language. Ship the fucking
documentation generator with the language. It's part of your job.