Progress on editable text control.
Basic editing works now but there are many things that still need work and
refinement. Yay for reinventing the biggest wheels, but I think this is one
we're really going to need.

This also includes a notion of keyboard focus for Root. Currently no
coordination is done between roots, but eventually that will be needed as well
because only one root can have focus at a time.