Separate Cursor into its own Element.
This harmonizes better with the philosophy that everything is broken down into
basic elements and then controls assemble those basic elements and control them
via user interactions. It also works around/solves the problem that controls
cannot themselves have styles.

The problem is that the control establishes a style context for its children,
but it cannot establish that same style context for itself. A style context has
to come from the parent of an element, not the element itself (it does not
exist at the time that we resolve the styles for it).

So a control can never be resolved with the style states that it defines for
its children. Short of a hacky terrible workaround for that, I'd rather just go
hard on the policy that controls do not render anything and thus need not have
styles. They always just contain child elements which do all the rendering.
Those elements will be properly styled based on the context set up by their
parent control.