Removed `map` from Source.
I want the reactive collections to also act as "sources" for their respective
non-reactive collection. So an RList can be observed in a fine-grained way,
hearing about additions, updates and removals, or it can be treated as a
Source<E[]> where you just want to know when anything changes and get the new
"list" as the changed value. Having `map` on Source makes that way more
complicated. So if you want to map, you have to have a more specific type than
Source. Source just lets you listen for values.