Various bits related to picking and the graph-based handling thereof.
Have SceneSystem take a reactive view of the current "touches" (mouse pointer
or touchpad data) and use them to raycast and update sets of hovered and
pressed entities maintained as components.  Added a node type that provides
a component value as a graph output so that we can use the hover state in the
graph (as well as a corresponding node type to set a component value).