Various UI reworking.
Made the element context contain a Model, a StyleContext and an ElementFactory.
Model is now an abstract class instead of just a POJO. You can create a
POJO-based model but we'll have other kinds of models as well. All this and a
few other tweaks are in preparation for List.

Also added notion of Action model elements, which are functions called by UI
components in response to user actions. Button now takes an Action instead of
an Emitter and an event. Emitter exposes a big API that isn't used for this
purpose, the button just needs to do something when it's clicked. Now it gets a
function that it calls, and that function can emit an event on an emitter if
desired.