Some react revamping.
Mainly try to keep things functions all the way down. A stream is just a
wrapper around an onEmit function. A mapped or filtered stream can just pass an
adapted listener to its underlying stream and then excuse itself from the
universe. This is not quite as so for subjects and values, but somewhat, so
wherever possible, we avoid capturing `this` and just close over the essential
things we need.

This also splits Subject and Value which I think is worth it. The main fallout
is needing to have/use Subject.constant(). Otherwise when one wants to abstract
over Subject or Value, one can use Source, which is better anyway because then
you could use Stream too if you wanted.