Added Mutable.bimap.
This allows you to do fancy things like:

  const foo = Mutable.local({bar: 1, baz: {quux: 3}})

and then:

  const bar = foo.bimap(f => f.bar, (f,bar) => ({...f, bar}))

which is now a mutable reactive value which reflects the value of the bar
property. You can call bar.update(3) and it will update foo with the value
{bar: 3, baz: {quux: 3}}.

You can project deeper, if desired:

  const quux = foo.bimap(f => f.baz.quux, (f,quux) => ({...f, {...f.bar, quux}}))

or chain bimaps:

  const quux = bar.bimap(b => b.quux, (b,quux) => ({...b, quux}))

though chaining creates & preserves the intermediate values, so you'd usually
only want to do that if you needed those for something.