Added RMap.keysSource & fixed bugs.
I opted not to add keysValue() because that would entail computing and sending
an old keys array along with the new one on every change, which almost no one
would want. So I expose it as a Source which under the hood is a Subject.

I also didn't want to call it keysSubject() because that would weirdly imply
that it might not resolve right away. So it's a source and the docs say that it
has a value right away.