A dynamic List.
OMG this involved a lot of thinking and rearranging of the plumbing, but in the
end the List component is now super simple.

I originally was going to have the data source for List be an RList, but
converting an RList of some values into an RList of Model instances (which
themselves contain reactive values) would be messy.

Also, especially in distributed/persistent situations, one rarely has their
data in a tidy little list. Databases or distributed data models tend to keep
collections of things as a map by some persistent id, and then you decide how
you want to order things based on some sort order. So reflecting that in the UI
model will probably make life easier down the road.

Also also, filtering an RList is messy, especially one that's mapped from raw
data to Models and I'm not sure I even want to support it. But with the
separation between "a map of all your data" and "show these specific keys in
this order" it because much easier to have your list display a filtered, or
differently sorted, version of the data.

Anyway, this is definitely a situation where I was able to spend the time to
write a really short letter.