Store both world and view positions/deltas in Hover.
When dragging objects, we want the world delta (which includes camera
movement, so that we can "hold" an object while moving around).  When dragging
the camera, don't want to include the camera movement, because that would
cause a feedback loop.  In that case, it's easier just to use view-relative
movement to translate in the local space of the camera.