Allow listener functions to request removal.
This allows us to fix issues with once() listeners. We were closing over the
remover function and calling them in the listener the first time it was called.
But for already ready values (constant subjects or values for example), the
listener would be called before the remover was returned.

We could do a bunch of fiddling to keep track of whether the listener was
called inline, but allowing listener functions to return a `Remove` sentinel is
simpler, adds useful functionality and is more efficient.