Factoring server bits out, create/resolve revamp.
Originally I had resolve return an async result, then I thought it would be
nice to have it just return a blank object that was populated with the real
object values once they arrived. Though this would be nice, this prevents us
from supporting "constant" object properties (which would be specified when the
object was created and then never change), and it introduced wonkiness in
access control because we had to be sure to resolve the object prior to doing
any access checks.

I've abandoned the cute idea and gone back to async resolution. This allows
objects to have constant properties and it allows the Subject to manage the
subscription lifetime. When no one is listening to the subject, we're not
subscribed. We could even use the stream of results nature of the Subject to
magically propagate new objects if the object migrated from one server to
another, but that's crazy and probably never going to happen.