Collided, tagged, entity id, log nodes.
Change deleteEntity to take entity id (or undefined) as input.

I am attempting to use Cannon.js's collision detection for non-dynamic bodies.
This may not be the best approach in the long term--for instance, I've already
found that Cannon's sweep-and-prune implementation tends to miss some
collisions--but it's something to start with.