Absolute layout, working on scroll view and graph viewer.
The absolute layout takes constraints consisting of position and (optionally)
size, also allowing children to stretch to fill the entire area.  It assumes
that children can overlap, so it processes mouse events in the reverse order
of rendering.

Rather than having scroll bars, the scroll view will support panning around
the contents with the mouse (and zooming with the mouse wheel).