Two test subjects walk into a room...
...and they post chat messages to one another. Yay.

I'm not sure I love how much manual message passing that we need. I would kill
for the ability to send closures between JavaScript VMs. Absent that, I may
define a standard protocol for making changes to some other object's data so
that we don't have a zillion one-off messages and handlers that just add data
to a map or update a property in a map value, etc.