Expose "connectedness" at a per-object level.
I'm still not sure exactly what I want to do here, but since different objects
come from different servers, it doesn't make sense to expose a global
"connected" state. On the other hand, keeping track of whether any individual
object is connected is kind of a PITA, so something else may be needed.

I'd also like to, if possible, keep objects around and then sync them back up
if we reconnect, so that we can potentially just sort of "freeze" if we lose
connection and then jump back up to currency when we reconnect and resubscribe.