Rethink! Simplify!
I had strayed too far from the virtual actor paradigm and was getting annoyed
with manual creation and async blah blah. This brings us back to a simpler
model with fewer failure modes.

Objects are not created or destroyed (it will eventually be possible to delete
persistent storage allocated for an object). I've thus given up dconst
properties. Just set values to what you want and don't change them. Database
rows don't have any "constant" values either (other than the key which is also
constant here), and we survive. If something never changes, yay, it never
changes.

Object keys are also always UUIDs. Screw numeric keys. If you want to "create"
an object, you invent a new key and you start talking to it. So now you know
what the key is at the moment you invent it. You don't have to wait for the
server to tell it to you via some error prone async round trip.

Objects are back to having a (resolution) state, which says whether they've
received their initial data or not. On the server you'll never observe an
object in the `resolving` state because we make sure an object is fully
resolved before we use it to process queue messages, but it's still a thing
internally. On the client, you can observe an object in the resolving state,
which is what you'll see until the server gets back to us with the initial
state of the object. That's fine, there's always going to be some async delay
while we get the object, and now you're back to being able to wire up all your
reactive listeners immediately after requesting the object, and they'll just
transition from the default values to the current object values when the object
is resolved.