Root size and positioning improvements.
1. You can configure a root to autoSize which will cause it to resize itself to
its preferred size whenever it's invalidated. This is good for dialogs that
might need to resize, or HUDs.

2. You can bind the origin of a root to an anchor point on the screen (given a
reactive view of the screen size). This will position the root properly when
either the root or screen size changes.

3. Root maintains its own origin vector now instead of having it separate. That
turned out to be a needless annoyance, and now Root wants to twiddle its own
origin anyway so it needs it.