Base cursor blink on focused-ness directly.
It was being based on state === "focused" but that's no longer precise enough
because now we can be hovered (and we may or may not be focused during that
time).

I was agonizing a bit when I added focus in the first place about whether it
should be a state, partly because I was thinking about hovered and how these
are two orthogonal pieces of information that we want to communicate. I may
change things later to have focused and hovered to "augment" the current state,
but I need to think about how that will be expressed in terms of style
configuration.

Ideally one would be able to, say, change the border when focused, and change
the background when hovered, and have both of those things show up when you
were both focused and hovered.