Don't let Node libs appear in client code.
This causes WebPack to think that it needs to include a bunch of Node emulation
crap, which we definitely do not want.

Instead we just inject the Node bits in data/server, which is not the greatest
place to do it, but at least it's code we know is exclusive to the server and
is highly likely to be included by any server stuff.