Support a repetition count when playing animations.
There's a lot that can be done here. The controls on
the AnimationAction are varied and complex. Rather than
adding configuration properties to try to support everything
we might want to do, perhaps we should have the node support
an onPlay() callback that receives the AnimationAction object
and can do whatever it likes.

I also wonder if we should add an output edge to the AnimationAction
node, that becomes true when the animation has finished playing?