Different reactive wiring for AnimationAction.
We map the mixer and the animation data into an action subject. Then we
separately join that with playing to start or stop the animation as desired.

Finally, we join action and playing and use that to create a subject which
completes with true when a finished event is emitted. Because the finished
subject is recreated when action or playing changes, it should properly reflect
both changing models/clips and manually stopped animations.