First pass at a Unity-like AnimationController (and corresponding node).
The configuration for the AnimationController consists of a set of named
states (including the initial "default" state and the special "any" state),
each associated with an optional animation.  States can opt to finish their
animations (that is, finish the number of repetitions specified) before
transitioning to a new state.  Transitions are controlled by a number of
named boolean conditions that become inputs on the AnimationControllerNode.
When a condition becomes true (and, if desired, the current animation ends),
the transition is executed.  The "any" state includes transitions that apply
to all states.  At the moment, you can't transition from a state to itself,
but you can transition to an intermediate state (perhaps without an
animation) and back.

Other things we'll want eventually will include some system of weighting or
probability associated with the transition (to give control over priority
and allow random variations), events in addition to conditions, etc.