Slight tweak: use promises to represent GLTFs in a (possible) state of loading.
Without this, we risk loading the same model multiple times if you subscribe
to the subject, unsubscribe, and resubscribe in rapid succession.  Three.js
doesn't provide an API for cancelling the download (unless maybe we use the
XMLHttpRequest instance provided to onProgress), and we wouldn't want to
cancel anyway; if you triggered the download once, then you probably want
the model cached for a while.