Fix for animations' not taking effect until update.
The root problem was that the animations were being started in
SceneSystem.render, when we update the hover states, which we do there because
the raycast relies on up-to-date scene graph transforms (i.e., it has to be
called after SceneSystem.update).  Unfortunately, starting an animation
doesn't immediately jump to the first frame, so the first frame would only
show after the call to AnimationSystem.update (in other words, after a frame
was already rendered using the default transforms).

To fix, we poke into the Three.js internals a little.  When we start an
action, we immediately update it to its first frame (using a deltaTime of 0
doesn't quite work, perhaps because there's a single-frame transition, so we
advance by 1/60 of a second).  Then we update the world matrices so that
they'll be ready for rendering, and then we reset the action time back to
zero in case that kind of single-frame timing matters for the animation.

Obviously it's not ideal to access "protected" members of the animation system
classes, but unfortunately the animation system isn't really designed for this
kind of subtle manipulation.