Revamp all the things! Rough, untested "indexes".
The main revamp is using paths everywhere to in the protocol messages and then
mapping paths to 16-bit unsigned integers under the hood in the protocol layer.
This allows us to avoid the manual assignment of oids to objects, which was
problematic for indices because objects that come from an index would have to
have their oid assigned by the server instead of the client. Problem solved.

Would have been easier except I first rewrote everything to use UUID instead
and then realized that UUID is not in fact globally unique, in spite of all the
u's in the name. It's only unique within a particular collection, and we *want*
to support using the same UUID as a key for "related" objects in different
collections. For example a ranch and its associated chat channel can use the
same UUID even though they are two different objects in different collections.
Otherwise you gotta keep track of a mapping from ranch UUID to chat UUID, and
who has time for that?

Anyway, indexes. Indexes are a way to have a view of some subset (or all) of
the objects in a collection. You define an index on an object that also has a
collection, and you can add where clause, and order by clauses. Then you
subscribe to that index and get a "view" of it. It effectively subscribes you
to *all* the distributed objects that match the index criteria. So you have
them all and you get change events for all of the matched objects. You also get
notifications when objects materialize into the index or become excluded from
it.

As you might expect, this is a Hard Problem (tm), and I am nowhere near having
solved it fully, but the basic groundwork has been laid and I might even have
"view all the objects in a collection" indexes kinda working, but I haven't
tested it. There's a *lot* of work left to do and dozens of thorny things to
sort out like whether to allow dynamic views with custom limits and start
positions, and what to do when objects are visible to the client via multiple
indices or via one or more indices and a direct subscription. And access
control. And a bunch of other stuff. But a few man years of effort down the
road and this is going to be super awesome.