Indexes are so yesterday. Tables and views are the new hotness.
After sleeping on it, I decided that there was a conceptual mismatch between a
fully operational distributed object (which is like a virtual actor: a
collection of data along with message handlers that can do things with that
data) and a table row (which is generally just a small collection of "dumb"
key/value pairs). You don't do the same things with them, and it would be
annoying to try to make a distributed object act like dumb data, and vice
versa.

So now we have a new concept: a dtable. It's kind of like a dcollection except
it stores dumb data (records), not fully operational distributed objects. You
can then define views of that table (which are equivalent to yesterday's
indices), and when you subscribe to a view, you get an RMap which materializes
that view.

Crucially, the table provides CRUD operations (CUD really because the R is
handled by the view, you can't just load a single table row), which is
something you do not do with distributed objects (which just "exist" because
that's how the virtual actor model works).

So essentially a dtable is like a dmap that you expect to get really big. So
instead of just having a whole map right inside the object, downloaded in full,
you have views of the table which you can subscribe to (or not), and which give
you some (ordered) subset of the full table.

There's still work to do to handle overlapping views efficiently, and
"extensible" views (for things like infinite scrolling where you request more
data when you get to the end of the data you have). We're also leaning very
heavily on Firebase for the implementation, but I'm not going to worry too much
about that until we care about using a different data store.