New approach to injecting props into children.
AKA How I Learnd to Stop Worrying and Love the Bomb.

This is an abomination, but the hoops I was trying to jump through to sneak
config hacks in during the element creation process were even more abominable.
(Hacks elsewhere, this is just preliminary support for other stuff.) At this
point, I'm throwing in the towel and just patching like the monkey I am. This
shit would look real good on that wall, I'm throwing it.