Have HTMLHost "shadow" text input fields.
It overlays an HTML input element atop the text field and allows it to handle
text input. This is necessary on mobile because we cannot listen to raw
keyboard events, but given our web nature, it's probably a good default
strategy. It enables us to leverage the many many years of effort that have
gone into text input handling on the native platforms.

There are some niggles where things don't line up perfectly on particular
platforms, and things shift a pixel or two when transitioning from overlaid
input field to internally handled statically rendered text. We can likely iron
those out eventually, but for now they're not too jarring.