Auth improvements.
We now store the last session auth info in local storage and reuse it
immediately for the next session. This avoids the whole "start as a guest, wait
to hear back from Firebase, then finally auth as who we are" delay. Now we auth
immediately with our last known user, if that info has expired, the server will
reject that auth, and when Firebase comes back with new auth, we'll try authing
again (or revert to guest-hood).

We also now defer processing messages from a client until it is authed, because
clients now start sessions with credentials that require async auth (i.e.
firebase).