Make spacer not consume all available space.
This was resulting in strange behavior that I don't think anyone really wants.
In unconstrained environments it made the spacer want to be as large as the
screen, and in general it just made it claim to want to be as big as the
available space.

For example, if a group said, "Hey, I have 100 pixels, how much do you want?" A
label might say 25, and a button says 30, and the spacer says 100. So the group
ends up thinking it wants 155 pixels of space. Not quite right. Layout proceeds
in two phases, first we compute how much space we want, then we compute how
much space to assign children based on how much space we got. So during the
"how much do you want" phase, a spacer should say 0 (or its minimum size).

With that behavior, during preferred size computation the label says 25, the
spacer says 5 (for a minimum gap, say) and the button says 30, and the group
"prefers" a size of 60. Then later when it is actually set to a size of 100
(because the group itself was stretched to fit a bigger space, say), it assigns
a size of 25 to the label, 30 to the button and 45 to the spacer (because it is
also annotated as "stretch"), which is what we want.

In changing this, I discovered and fixed a couple of things.

1. Menu items contained a spacer that was "preferring" the default width of the
menu, which meant that the space between menu item and shortcut would be
whatever space was initially provided to the entire menu popup as its hint
width (which was coming from DropdownStyle.minWidth which defaulted to 100). So
the spacers would claim a preferred width of 100 and then the label and
shortcut would be added on to that, thus making the spacers 100 through a
rather convoluted path. Now the spacers in menu items are just set to a
specific min width. They provide at least that much space, and are also
assigned any extra space for menu items with shorter text than the widest menu
item.

With this behavior, the minWidth configuration for Dropdown should be
unnecessary, at least in terms of making spacers behave. You still might want a
min width if you don't want a menu to be really skinny if it had only like "A",
"B" and "C" as menu items, or something. Though I'd like to eventually support
that via some universal minimum size mechanism so that any component could
clamp its preferred size in a dimension to some specific minimum value.

2. GraphViewer has a menu bar across its top that was taking advantage of
spacers "prefer to be the width of the entire screen" behavior to mask the fact
that its desired "offPolicy: stretch" config override was not happening thanks
to the "idiosyncracies" of constructors and inheritance.

GraphViewer's constructor was initially:

  constructor (readonly ctx :ElementContext, parent :Element, readonly config :GraphViewerConfig) {
    super(ctx, parent, {...config, offPolicy: "stretch"})

but the unadjusted config is what gets saved and not the adjusted one passed to
the super constructor. In this case the super class didn't even save the config
that was passed to it, but even if it had, it would have been overridden by the
property defined by GraphViewer (which contained the unadjusted config value).
I was just ranting about this exact sort of problem yesterday. Fancy that.

Anyway the graph viewer was not properly stretching out the menu bar at its top
because that stretch config was being ignored.