New new approach to subelement injection.
The main motivation is that I want elements to be able to fail if they are
improperly configured, but the way we were doing injection previously required
them to be lenient about missing configuration so that it could be injected
later. But that makes every single _normal_ use of the element into a "silent,
hard to debug failure" opportunity instead of a "loud immediate failure with
helpful feedback that tells you exactly where you screwed up" opportunity. We
want the latter.

So now we are back to injecting things into the subelement configuration by way
of refining the ElementContext that is used to create them. This is also
combined with some function default value sneakiness to allow creation of
properties prior to calling the super constructor.

The final upshot of this is that `label` can now fail again if you either pass
in a bogus model path or forget to supply a text property entirely, instead of
just silently displaying nothing.