Added reactive views of list/set/map size.
Also changed RMap.keysView to return an Iterable rather than an
IterableIterator. The latter is still just an iterator, and if you have
multiple listeners to that value, the first one will use up the iterator and
the rest will think it's done.

This was previously being masked by the fact that every call to keysValue()
returned a new derived Value, but it was still possible to screw yourself by
sticking the result of keysValue into a local and adding multiple listeners to
it. Now that we use poor man's lazy properties to only ever create one
keysValue instance, the problem jumped right to the surface.