Auto-reconnect client after disconnect.
This also simplifies the client: we only have one server so we'll only support
one server at the moment.

Later when we have multiple servers and I have a clearer idea of how I want to
do object routing (figuring out which server hosts a particular object), I'll
reinstitute the complexity of connections to multiple servers and handle the
additional complexity of dealing with the fact that an object might move from
one server to another while we were disconnected.