General purpose RPC services over channels.
This will evolve a bit, but it's functionality that we're going to want.

The design relies on a bit more magic than I would like. In an ideal world,
you'd declare:

@rservice
interface MyService {

 @rcall("string", "int32")
 someMethod (arg1 :string, arg2 :number) :void
}

but JavaScript annotations do not work on interfaces. Nor do they work on
abstract methods. So instead you have to declare a concrete class (which I call
the "protocol") that stubs out all the methods. Meh.

Then I have to do some type magic (thank Anders for TypeScript's crazier type
system features) to "extract" an interface from the protocol class, and that is
what you implement in your service providers and what defines the type of the
proxy you get on the client.