Fix issues with mouse clicks going to multiple roots.
- Dispatch the event to roots in the reverse of the draw order.
- Stop once the event's cancelBubble has been set.

Probably this needs to be done with the other event handlers too,
I'll leave it as an exercise for mdb. :)

MDN seems to indicate that cancelBubble is old badness and the new
goodness is stopPropagation() but I don't know how to check that
stopPropagation() has been called. And we just set cancelBubble.