Added ComponentObserver & tweaked sparse component default.
ComponentObserver allows one to observe all changes to all entitys' values.

SparseValueComponent no longer creates a map entry for the default value if an
entity starts at the default value, it just leaves the map empty. This works
better with the "delete the mapping when updated to undefined" behavior, in
that a component reverts back to the default value when not configured with its
own value. Though this may not be desired in all use cases, it is now
documented, so if someone wants something special they can use a different
component.

The Component.getValue support was factored out into a lazily added component
observer. Additionally two changes were made there: one was to check changes
against `eq` and only emit on actual change (as that's what Values promise in
their contract), and the other was to keep created values around forever.

Otherwise it would be possible to call getValue(), listen to the value and then
unlisten to it (causing it to clear out both _values and _changeFns), but keep
that Value object around. Then call getValue() again which would create a new
value, subscribe to the new value to cause it to stick its changeFn in
_changeFns, then finally subscribe to the _old_ value again, causing it to
stick its changeFn into _changeFns, overwriting the second value's changeFn.
Though this would be a pretty obscure failure mode, it's pretty cheap to just
keep the value around and avoid potential weirdness if someone does something
crazy.