Properly skip disabled entities.
Also allowed DenseValueComponent to take no default value and mandate that all
entities specify an initial value in their configuration. This could be useful
for SparseValueComponent, but interaction with deleting the current value to
revert to the default becomes tricky, We may want two sparse value components,
one that uses a default and one that doesn't.