Queue fix & improvement.
When posting new messages from the context supplied to a queue handler, we need
to switch to system auth, not keep using the auth of the original requester.

Added a `system` property for queues which says all messages to this queue
should come from the system. Simplifies access checking for the auth handler
and encourages users to separate user queues from system queues. Messages to
user "writable" queues should be carefully access checked, messages to system
queues can be less paranoid.