Revalidate/rerender in setSize().
Otherwise we compute our preferred size and then spend one frame invalid, so
any changes to the UI do not trigger a recomputation of preferred size and when
we validate on the next frame we can end up using stale preferred size data.

Now we immediately revalidate and rerender, and if things change on the next
frame, we respond to that normally with another revalidation.