A new reactive primitive: Buffer.
This is like a Value but which is designed around mutate-in-place semantices
instead "value" semantics. Its API thus accommodates users who want to jam a
struct or an array into it and then update those things in place and notify
listeners of the changes. It does not (due to mutation in place) notify
listeners of changes because old values are implicitly destroyed by the in
place updates.

This reverts the addition of an update merger to Mutable because that was a
dangerous way to support this usage pattern. Buffer does a better job of it and
does not make promises that it cannot keep.